test/medicalassistant0824/requirements.txt
Flask
gunicorn
line-bot-sdk
APScheduler
psycopg2